修改员工角色等entity，mapper等

+++ user-api/src/main/java/com/jaagro/user/api/dto/request/UpdateEmpRoleDto.java
@@ -15,16 +15,16 @@ public class UpdateEmpRoleDto implements Serializable {
     /**
      * 员工角色表ID
      */
-    private Long id;
+    private Integer id;
 
     /**
      * 员工id
      */
-    private Long employeeId;
+    private Integer employeeId;
 
     /**
      * 角色id
      */
-    private Long roleId;
+    private Integer roleId;
 
 }
