+++ user-biz/src/main/java/com/jaagro/user/biz/mapper/EmployeeCooperationMapper.java
@@ -9,32 +9,38 @@ import java.util.List;
  */
 public interface EmployeeCooperationMapper {
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insert(EmployeeCooperation record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insertSelective(EmployeeCooperation record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    EmployeeCooperation selectByPrimaryKey(Long id);
+    EmployeeCooperation selectByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKeySelective(EmployeeCooperation record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKey(EmployeeCooperation record);
