+++ user-biz/src/main/java/com/jaagro/user/biz/mapper/EmployeeMapper.java
@@ -9,32 +9,38 @@ import java.util.List;
 
 public interface EmployeeMapper {
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insert(Employee record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insertSelective(Employee record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    Employee selectByPrimaryKey(Long id);
+    Employee selectByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKeySelective(Employee record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKey(Employee record);
 
@@ -60,7 +66,7 @@ public interface EmployeeMapper {
      * @param id
      * @return
      */
-    UserInfo getUserInfoById(Long id);
+    UserInfo getUserInfoById(Integer id);
 
     /**
      * 判断用户名手机号是否已存在
