+++ user-biz/src/main/java/com/jaagro/user/biz/mapper/PermissionMapper.java
@@ -6,32 +6,38 @@ import java.util.List;
 
 public interface PermissionMapper {
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insert(Permission record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int insertSelective(Permission record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
-    Permission selectByPrimaryKey(Long id);
+    Permission selectByPrimaryKey(Integer id);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKeySelective(Permission record);
 
     /**
-     * @mbggenerated 2018-08-18
+     *
+     * @mbggenerated 2018-08-22
      */
     int updateByPrimaryKey(Permission record);
